+++ Cargo.toml
@@ -16,7 +16,6 @@ members = [
   "custody",
   "measure",
   "storage",
-  "storage2",
   "tct",
   "tct-property-test",
   "tct-visualize",
