+++ component/src/app/mod.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use penumbra_chain::params::FmdParameters;
 use penumbra_chain::{genesis, StateWriteExt as _};
-use penumbra_storage2::{AppHash, State, StateTransaction, StateWrite, Storage};
+use penumbra_storage::{AppHash, State, StateTransaction, StateWrite, Storage};
 use penumbra_transaction::Transaction;
 use tendermint::abci::{self, types::ValidatorUpdate};
 use tracing::instrument;
