+++ component/src/governance/component.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use anyhow::Result;
 use async_trait::async_trait;
 use penumbra_chain::genesis;
-use penumbra_storage2::{State, StateTransaction};
+use penumbra_storage::{State, StateTransaction};
 use penumbra_transaction::Transaction;
 use tendermint::abci;
 use tracing::instrument;
@@ -20,7 +20,7 @@ impl Component for Governance {
     //#[instrument(name = "governance", skip(state, _app_state))]
     async fn init_chain(state: &mut StateTransaction, _app_state: &genesis::AppState) {
         // Initialize the unfinished proposals tracking key in the JMT.
-        // TODO: Replace with the new range queries in storage2
+        // TODO: Replace with the new range queries in storage
         state
             .put_unfinished_proposals(ProposalList::default())
             .await;
