+++ component/src/governance/tally.rs
@@ -6,7 +6,7 @@ use penumbra_transaction::action::Vote;
 use super::{proposal::Withdrawn, StateReadExt as _};
 use crate::stake::StateReadExt as _;
 use penumbra_chain::StateReadExt as _;
-use penumbra_storage2::StateRead;
+use penumbra_storage::StateRead;
 
 use super::proposal::Outcome;
