+++ component/src/ibc/client.rs
@@ -145,7 +145,7 @@ pub fn validate_penumbra_client_state(
     }
 
     // check client proof specs match penumbra proof specs
-    if penumbra_storage2::PENUMBRA_PROOF_SPECS.clone() != tm_client_state.proof_specs {
+    if penumbra_storage::PENUMBRA_PROOF_SPECS.clone() != tm_client_state.proof_specs {
         return Err(anyhow::anyhow!(
             "invalid client state: proof specs do not match"
         ));
