+++ component/src/shielded_pool/component.rs
@@ -11,7 +11,7 @@ use penumbra_chain::{
     Epoch, NoteSource, StateReadExt as _,
 };
 use penumbra_crypto::{asset, note, IdentityKey, NotePayload, Nullifier, Value};
-use penumbra_storage2::{State, StateRead, StateTransaction, StateWrite};
+use penumbra_storage::{State, StateRead, StateTransaction, StateWrite};
 use penumbra_tct as tct;
 use penumbra_transaction::{action::swap_claim::List as SwapClaimBodyList, Action, Transaction};
 use tct::Tree;
