+++ component/src/shielded_pool/note_manager.rs
@@ -4,7 +4,7 @@ use async_trait::async_trait;
 use decaf377::{Fq, Fr};
 use penumbra_chain::{sync::AnnotatedNotePayload, NoteSource};
 use penumbra_crypto::{ka, Address, Note, NotePayload, Nullifier, One, Value};
-use penumbra_storage2::StateWrite;
+use penumbra_storage::StateWrite;
 use penumbra_tct as tct;
 use tracing::instrument;
