+++ component/src/shielded_pool/supply.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use async_trait::async_trait;
 use penumbra_chain::KnownAssets;
 use penumbra_crypto::asset::{self, Asset, Denom};
-use penumbra_storage2::{StateRead, StateWrite};
+use penumbra_storage::{StateRead, StateWrite};
 use tracing::instrument;
 
 use super::state_key;
