+++ storage/src/state/read.rs
@@ -10,7 +10,6 @@ use penumbra_proto::{Message, Protobuf};
 /// Read access to chain state.
 // This needs to be a trait because we want to implement it over both `State` and `StateTransaction`,
 // mainly to support RPC methods.
-//#[async_trait(?Send)]
 #[async_trait]
 pub trait StateRead: Send + Sync {
     /// Gets a value from the verifiable key-value store as raw bytes.
@@ -152,17 +151,6 @@ pub trait StateRead: Send + Sync {
     /// TODO: rename to `ephemeral_get` ?
     /// TODO: should this be `&'static str`?
     fn get_ephemeral<T: Any + Send + Sync>(&self, key: &str) -> Option<&T>;
-
-    // TODO: remove
-    /*
-    /// Retrieve all objects for keys matching a prefix from the ephemeral key-value store.
-    ///
-    /// TODO: rename to `ephemeral_prefix` ?
-    fn prefix_ephemeral<'a, T: Any + Send + Sync>(
-        &'a self,
-        prefix: &'a str,
-    ) -> Box<dyn Iterator<Item = (&'a str, &'a T)> + 'a>;
-    */
 }
 
 // Merge a RYW cache iterator with a backend storage stream to produce a new Stream,
@@ -275,18 +263,8 @@ impl<'a, S: StateRead + Send + Sync> StateRead for &'a S {
     fn get_ephemeral<T: Any + Send + Sync>(&self, key: &str) -> Option<&T> {
         (**self).get_ephemeral(key)
     }
-
-    /*
-    fn prefix_ephemeral<'b, T: Any + Send + Sync>(
-        &'b self,
-        prefix: &'b str,
-    ) -> Box<dyn Iterator<Item = (&'b str, &'b T)> + 'b> {
-        (**self).prefix_ephemeral(prefix)
-    }
-    */
 }
 
-//#[async_trait(?Send)]
 #[async_trait]
 impl<'a, S: StateRead + Send + Sync> StateRead for &'a mut S {
     async fn get_raw(&self, key: &str) -> Result<Option<Vec<u8>>> {
@@ -307,13 +285,4 @@ impl<'a, S: StateRead + Send + Sync> StateRead for &'a mut S {
     fn get_ephemeral<T: Any + Send + Sync>(&self, key: &str) -> Option<&T> {
         (**self).get_ephemeral(key)
     }
-
-    /*
-    fn prefix_ephemeral<'b, T: Any + Send + Sync>(
-        &'b self,
-        prefix: &'b str,
-    ) -> Box<dyn Iterator<Item = (&'b str, &'b T)> + 'b> {
-        (**self).prefix_ephemeral(prefix)
-    }
-    */
 }
