--- storage2/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "penumbra-storage2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-penumbra-proto = { path = "../proto" }
-penumbra-crypto = { path = "../crypto" }
-penumbra-tct = { path = "../tct" }
-
-jmt = { git = "https://github.com/penumbra-zone/jellyfish-merkle.git", branch = "newmain" }
-ibc-proto = { git = "https://github.com/penumbra-zone/ibc-rs", branch = "penumbra-034" }
-tokio = { version = "1.21.1", features = ["full", "tracing"] }
-tokio-stream = { version = "0.1.11" }
-sha2 = "0.9"
-tempfile = "3.3.0"
-ibc = { git = "https://github.com/penumbra-zone/ibc-rs", branch = "penumbra-034" }
-anyhow = "1"
-once_cell = "1.8"
-bincode = "1.3"
-tendermint = { git = "https://github.com/penumbra-zone/tendermint-rs", branch = "penumbra-034" }
-async-trait = "0.1.52"
-async-stream = "0.3.3"
-tracing = "0.1"
-rocksdb = "0.19.0"
-futures = "0.3"
-tonic = "0.8.1"
-hex = "0.4"
-metrics = "0.19.0"
-parking_lot = "0.12"
-ics23 = { git = "https://github.com/penumbra-zone/ics23", branch = "penumbra-034" }
-
-[dev-dependencies]
-tempfile = "3.3.0"
-tracing-subscriber = "0.3"
